# lib/api-zod/src/generated/types/promptInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Prompt Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface PromptInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  content: string;
  description?: string;
  categoryId: number;
  tags?: string[];
  authorUsername: string;
  isPublic?: boolean;
}