# lib/api-zod/src/generated/types/promptListResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Prompt Marketplace API
 * OpenAPI spec version: 0.1.0
 */
import type { Prompt } from './prompt';

export interface PromptListResponse {
  prompts: Prompt[];
  total: number;
}