# lib/api-zod/src/generated/types/saveToggleResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Prompt Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface SaveToggleResponse {
  saved: boolean;
  saveCount: number;
}