# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Prompt Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  /** @minLength 1 */
  displayName?: string;
  bio?: string;
  avatarUrl?: string;
  categories?: string[];
}